fetch/metadata/form.tentative.https.sub.html

<!DOCTYPE html>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/resources/testdriver.js></script>
<script src=/resources/testdriver-vendor.js></script>
<script src=/fetch/metadata/resources/helper.js></script>
<script src=/common/utils.js></script>
<body>
<script>
  const USER = true;
  const FORCED = false;

  function create_test(host, user_activated, expectations) {
    async_test(t => {
      let nonce = token();
      let i = document.createElement('form');

      window.addEventListener('message', t.step_func(e => {
        assert_header_equals(e.data, expectations, `{{host}} -> ${host} iframe: ${user_activated ? "user-activated" : "forced"}`);
        t.done();
      }));

      let url = `https://${host}/fetch/metadata/resources/post-to-owner.py`;
      i.action = url;
      i.id = "form" + nonce;
      i.target = "_blank";

      if (user_activated == FORCED) {
        document.body.appendChild(i);
        let j = i.submit();
      } else if (user_activated == USER) {
        let s = document.createElement('input');
        s.type = "submit";
        s.id = "submit" + nonce;
        i.appendChild(s);
        document.body.appendChild(i);

        let k = test_driver.click(s);
      }
      // t.add_cleanup(_ => iw.close());
    }, `{{host}} -> ${host} iframe: ${user_activated ? "user-activated" : "forced"}`);
  }

  create_test("{{host}}:{{ports[https][0]}}", FORCED, {
    "site": "same-origin",
    "user": "",
    "mode": "navigate",
    "dest": "document"
  });

  create_test("{{hosts[][www]}}:{{ports[https][0]}}", FORCED, {
    "site": "same-site",
    "user": "",
    "mode": "navigate",
    "dest": "document"
  });

  create_test("{{hosts[alt][www]}}:{{ports[https][0]}}", FORCED, {
    "site": "cross-site",
    "user": "",
    "mode": "navigate",
    "dest": "document"
  });

  create_test("{{host}}:{{ports[https][0]}}", USER, {
    "site": "same-origin",
    "user": "?1",
    "mode": "navigate",
    "dest": "document"
  });

  create_test("{{hosts[][www]}}:{{ports[https][0]}}", USER, {
    "site": "same-site",
    "user": "?1",
    "mode": "navigate",
    "dest": "document"
  });

  create_test("{{hosts[alt][www]}}:{{ports[https][0]}}", USER, {
    "site": "cross-site",
    "user": "?1",
    "mode": "navigate",
    "dest": "document"
  });
</script>
